ci: upload rock as artefact and do not fail on trivy scans (#66) (#67)
* ci: upload rock as artefact and do not fail on trivy scans

This commit ensures the sample rock in this repository is not published to the
GHCR registry as it is not supported by the team, its only purpose is to serve
as an example and to be used in tests.
This commit also prevents the automatic trivy scans to exit with status code 1,
which was causing the CI to fail. While it is okay to scan the image, since this
is not a production or supported rock, we don't mind too much about the scan results.

Fixes #49